requirements.txt
configargparse
pyyaml
scipy
scikit-image
trimesh
tensorboard
tqdm
sklearn
plyfile